Split prosopography out of the IPC monolith
First domain extracted, and the template for the rest. Three contiguous blocks
move together so all three hot files shrink in the same change:

  electron/ipc.ts     1332-1389 -> electron/ipc/prosopography.ts     (53 handlers)
  electron/preload.ts   62-114  -> electron/preload/prosopography.ts (53 bindings)
  shared/types.ts     7467-7520 -> shared/api/prosopography.ts       (53 signatures)

registerIpc was a single 6,000-line function, and what made it resist being cut
into files was the state its handlers closed over. Measured first, that state
turned out to be almost entirely domain-local — the four AbortController maps,
the toolkit and translate signal maps, the presenter's pending imports all belong
to one domain each. Only two things are genuinely shared, and electron/ipc/context.ts
carries exactly those: the `h` wrapper, so every channel keeps the same payload
and error localization, and the main-window accessor.

Nothing about the public shape changes. window.nodus stays flat, NodusApi extends
the slice, no channel is renamed, and the 53 prosopography symbols were verified
to be used nowhere else in ipc.ts before moving.

Two prosopography tests asserted on IPC by reading electron/ipc.ts and
electron/preload.ts as text and matching a channel string, so they broke on a
move that changed nothing about the wiring. They now ask the channel census
whether each channel is registered AND reachable — a stronger assertion than the
old substring match, and one that survives the next move. Sixty test scripts read
those files as text; the rest are left alone until a later extraction touches them.
