+++ electron/preload.ts
@@ -11,6 +11,7 @@ import type {
   ChapterRelationsProgress,
 } from '@shared/types';
 import type { PresenterAction as PresenterControlAction } from '@shared/presenterState';
+import { prosopographyApi } from './preload/prosopography';
 
 // Tracks the research-chat stream currently in flight so `cancelResearchChat`
 // can abort it without the renderer having to juggle request ids. Only one chat
@@ -55,63 +56,15 @@ function readInitialOverlayPlacement(): NodiOverlayPlacement {
 }
 
 // Minimal, typed surface exposed to the renderer. No Node, no direct IPC names leak.
+//
+// Domain slices live in electron/preload/ and are spread in below; `NodusApi` on
+// this object is what makes a method lost during a move a typecheck failure
+// rather than an `undefined` the renderer discovers at runtime.
 const api: NodusApi = {
+  ...prosopographyApi,
   getSettings: () => ipcRenderer.invoke('settings:get'),
   updateSettings: (patch) => ipcRenderer.invoke('settings:update', patch),
   listVaults: () => ipcRenderer.invoke('vaults:list'),
-  getProsopPopulationWorkspace: () => ipcRenderer.invoke('prosop:population:workspace'),
-  updateProsopStudy: (patch) => ipcRenderer.invoke('prosop:study:update', patch),
-  createProsopMethodologyDraft: (actor) => ipcRenderer.invoke('prosop:methodology:createDraft', actor),
-  replaceProsopCriteria: (versionId, criteria) => ipcRenderer.invoke('prosop:methodology:replaceCriteria', versionId, criteria),
-  publishProsopMethodology: (versionId, changeSummary, actor) => ipcRenderer.invoke('prosop:methodology:publish', versionId, changeSummary, actor),
-  createProsopQuestionnaireDraft: (input) => ipcRenderer.invoke('prosop:questionnaire:createDraft', input),
-  saveProsopVariableRevision: (questionnaireVersionId, input) => ipcRenderer.invoke('prosop:questionnaire:saveVariable', questionnaireVersionId, input),
-  deleteProsopVariableRevision: (questionnaireVersionId, variableId) => ipcRenderer.invoke('prosop:questionnaire:deleteVariable', questionnaireVersionId, variableId).then(() => undefined),
-  publishProsopQuestionnaire: (questionnaireVersionId, changeSummary, actor) => ipcRenderer.invoke('prosop:questionnaire:publish', questionnaireVersionId, changeSummary, actor),
-  saveProsopVocabulary: (input) => ipcRenderer.invoke('prosop:vocabulary:save', input),
-  saveProsopVocabularyTerm: (input) => ipcRenderer.invoke('prosop:vocabulary:saveTerm', input),
-  getProsopSourcesWorkspace: () => ipcRenderer.invoke('prosop:sources:workspace'),
-  saveProsopSource: (input) => ipcRenderer.invoke('prosop:sources:save', input),
-  deleteProsopSource: (sourceId) => ipcRenderer.invoke('prosop:sources:delete', sourceId).then(() => undefined),
-  saveProsopSourceSegment: (input) => ipcRenderer.invoke('prosop:sources:saveSegment', input),
-  saveProsopCaptureTemplate: (input) => ipcRenderer.invoke('prosop:capture:saveTemplate', input),
-  importProsopDelimited: (input) => ipcRenderer.invoke('prosop:capture:importDelimited', input),
-  reviewProsopCaptureRow: (captureRowId, status) => ipcRenderer.invoke('prosop:capture:reviewRow', captureRowId, status),
-  getProsopObservationsWorkspace: () => ipcRenderer.invoke('prosop:observations:workspace'),
-  saveProsopFactoid: (input) => ipcRenderer.invoke('prosop:factoids:save', input),
-  reviewProsopFactoid: (factoidId, status, reviewedBy) => ipcRenderer.invoke('prosop:factoids:review', factoidId, status, reviewedBy),
-  saveProsopMissingValue: (input) => ipcRenderer.invoke('prosop:missing:save', input),
-  saveProsopResolution: (input) => ipcRenderer.invoke('prosop:resolutions:save', input),
-  retireProsopResolution: (resolutionId) => ipcRenderer.invoke('prosop:resolutions:retire', resolutionId).then(() => undefined),
-  getProsopIdentityWorkspace: () => ipcRenderer.invoke('prosop:identity:workspace'),
-  createProsopPerson: (input) => ipcRenderer.invoke('prosop:identity:createPerson', input),
-  saveProsopNameAttestation: (input) => ipcRenderer.invoke('prosop:identity:saveAttestation', input),
-  searchProsopIdentityCandidates: (literalName, limit) => ipcRenderer.invoke('prosop:identity:candidates', literalName, limit),
-  saveProsopIdentityHypothesis: (input) => ipcRenderer.invoke('prosop:identity:saveHypothesis', input),
-  decideProsopIdentityHypothesis: (hypothesisId, status, reviewedBy) => ipcRenderer.invoke('prosop:identity:decideHypothesis', hypothesisId, status, reviewedBy),
-  mergeProsopPersons: (survivorId, absorbedId, rationale, actor) => ipcRenderer.invoke('prosop:identity:merge', survivorId, absorbedId, rationale, actor),
-  reverseProsopPersonMerge: (mergeId, actor) => ipcRenderer.invoke('prosop:identity:reverseMerge', mergeId, actor).then(() => undefined),
-  saveProsopAuthorityId: (input) => ipcRenderer.invoke('prosop:identity:saveAuthority', input),
-  saveProsopOrganization: (input) => ipcRenderer.invoke('prosop:identity:saveOrganization', input),
-  getProsopMembershipWorkspace: () => ipcRenderer.invoke('prosop:membership:workspace'),
-  saveProsopMembership: (input) => ipcRenderer.invoke('prosop:membership:save', input),
-  saveProsopCohort: (input) => ipcRenderer.invoke('prosop:cohorts:save', input),
-  refreshProsopDynamicCohort: (cohortId) => ipcRenderer.invoke('prosop:cohorts:refresh', cohortId),
-  getProsopAnalysisWorkspace: () => ipcRenderer.invoke('prosop:analysis:workspace'),
-  runProsopAnalysis: (input) => ipcRenderer.invoke('prosop:analysis:run', input),
-  getProsopNetworksWorkspace: () => ipcRenderer.invoke('prosop:networks:workspace'),
-  saveProsopNetworkLayer: (input) => ipcRenderer.invoke('prosop:networks:saveLayer', input),
-  saveProsopNetworkEdge: (input) => ipcRenderer.invoke('prosop:networks:saveEdge', input),
-  deriveProsopCooccurrenceLayer: (layerId) => ipcRenderer.invoke('prosop:networks:deriveCooccurrence', layerId),
-  searchProsopography: (query, kind) => ipcRenderer.invoke('prosop:search', query, kind),
-  listProsopProposals: () => ipcRenderer.invoke('prosop:proposals:list'),
-  createProsopProposal: (input) => ipcRenderer.invoke('prosop:proposals:create', input),
-  decideProsopProposal: (proposalId, status, reviewedBy, decisionNote) => ipcRenderer.invoke('prosop:proposals:decide', proposalId, status, reviewedBy, decisionNote),
-  saveProsopNoteLink: (input) => ipcRenderer.invoke('prosop:notes:link', input),
-  exportProsopLongRows: () => ipcRenderer.invoke('prosop:export:long'),
-  exportProsopIpif: () => ipcRenderer.invoke('prosop:export:ipif'),
-  auditProsopIntegrity: () => ipcRenderer.invoke('prosop:integrity:audit'),
-  seedProsopDemo: () => ipcRenderer.invoke('prosop:demo:seed'),
   // Nodi companion: notifications
   listNotifications: () => ipcRenderer.invoke('nodi:notifications:list'),
   markNotificationsRead: () => ipcRenderer.invoke('nodi:notifications:markRead'),
