+++ electron/preload/prosopography.ts
@@ -1,0 +1,64 @@
+// Prosopography half of the renderer bridge, paired with electron/ipc/prosopography.ts.
+//
+// Typed as ProsopographyApi rather than as a loose object: the compiler is what
+// guarantees the slice stays complete, so a method dropped while moving it here
+// fails `npm run typecheck` instead of reaching a user. The channel names on the
+// other side of each call are checked by scripts/test-ipc-contract.mjs.
+import { ipcRenderer } from 'electron';
+import type { ProsopographyApi } from '@shared/api/prosopography';
+
+export const prosopographyApi: ProsopographyApi = {
+  getProsopPopulationWorkspace: () => ipcRenderer.invoke('prosop:population:workspace'),
+  updateProsopStudy: (patch) => ipcRenderer.invoke('prosop:study:update', patch),
+  createProsopMethodologyDraft: (actor) => ipcRenderer.invoke('prosop:methodology:createDraft', actor),
+  replaceProsopCriteria: (versionId, criteria) => ipcRenderer.invoke('prosop:methodology:replaceCriteria', versionId, criteria),
+  publishProsopMethodology: (versionId, changeSummary, actor) => ipcRenderer.invoke('prosop:methodology:publish', versionId, changeSummary, actor),
+  createProsopQuestionnaireDraft: (input) => ipcRenderer.invoke('prosop:questionnaire:createDraft', input),
+  saveProsopVariableRevision: (questionnaireVersionId, input) => ipcRenderer.invoke('prosop:questionnaire:saveVariable', questionnaireVersionId, input),
+  deleteProsopVariableRevision: (questionnaireVersionId, variableId) => ipcRenderer.invoke('prosop:questionnaire:deleteVariable', questionnaireVersionId, variableId).then(() => undefined),
+  publishProsopQuestionnaire: (questionnaireVersionId, changeSummary, actor) => ipcRenderer.invoke('prosop:questionnaire:publish', questionnaireVersionId, changeSummary, actor),
+  saveProsopVocabulary: (input) => ipcRenderer.invoke('prosop:vocabulary:save', input),
+  saveProsopVocabularyTerm: (input) => ipcRenderer.invoke('prosop:vocabulary:saveTerm', input),
+  getProsopSourcesWorkspace: () => ipcRenderer.invoke('prosop:sources:workspace'),
+  saveProsopSource: (input) => ipcRenderer.invoke('prosop:sources:save', input),
+  deleteProsopSource: (sourceId) => ipcRenderer.invoke('prosop:sources:delete', sourceId).then(() => undefined),
+  saveProsopSourceSegment: (input) => ipcRenderer.invoke('prosop:sources:saveSegment', input),
+  saveProsopCaptureTemplate: (input) => ipcRenderer.invoke('prosop:capture:saveTemplate', input),
+  importProsopDelimited: (input) => ipcRenderer.invoke('prosop:capture:importDelimited', input),
+  reviewProsopCaptureRow: (captureRowId, status) => ipcRenderer.invoke('prosop:capture:reviewRow', captureRowId, status),
+  getProsopObservationsWorkspace: () => ipcRenderer.invoke('prosop:observations:workspace'),
+  saveProsopFactoid: (input) => ipcRenderer.invoke('prosop:factoids:save', input),
+  reviewProsopFactoid: (factoidId, status, reviewedBy) => ipcRenderer.invoke('prosop:factoids:review', factoidId, status, reviewedBy),
+  saveProsopMissingValue: (input) => ipcRenderer.invoke('prosop:missing:save', input),
+  saveProsopResolution: (input) => ipcRenderer.invoke('prosop:resolutions:save', input),
+  retireProsopResolution: (resolutionId) => ipcRenderer.invoke('prosop:resolutions:retire', resolutionId).then(() => undefined),
+  getProsopIdentityWorkspace: () => ipcRenderer.invoke('prosop:identity:workspace'),
+  createProsopPerson: (input) => ipcRenderer.invoke('prosop:identity:createPerson', input),
+  saveProsopNameAttestation: (input) => ipcRenderer.invoke('prosop:identity:saveAttestation', input),
+  searchProsopIdentityCandidates: (literalName, limit) => ipcRenderer.invoke('prosop:identity:candidates', literalName, limit),
+  saveProsopIdentityHypothesis: (input) => ipcRenderer.invoke('prosop:identity:saveHypothesis', input),
+  decideProsopIdentityHypothesis: (hypothesisId, status, reviewedBy) => ipcRenderer.invoke('prosop:identity:decideHypothesis', hypothesisId, status, reviewedBy),
+  mergeProsopPersons: (survivorId, absorbedId, rationale, actor) => ipcRenderer.invoke('prosop:identity:merge', survivorId, absorbedId, rationale, actor),
+  reverseProsopPersonMerge: (mergeId, actor) => ipcRenderer.invoke('prosop:identity:reverseMerge', mergeId, actor).then(() => undefined),
+  saveProsopAuthorityId: (input) => ipcRenderer.invoke('prosop:identity:saveAuthority', input),
+  saveProsopOrganization: (input) => ipcRenderer.invoke('prosop:identity:saveOrganization', input),
+  getProsopMembershipWorkspace: () => ipcRenderer.invoke('prosop:membership:workspace'),
+  saveProsopMembership: (input) => ipcRenderer.invoke('prosop:membership:save', input),
+  saveProsopCohort: (input) => ipcRenderer.invoke('prosop:cohorts:save', input),
+  refreshProsopDynamicCohort: (cohortId) => ipcRenderer.invoke('prosop:cohorts:refresh', cohortId),
+  getProsopAnalysisWorkspace: () => ipcRenderer.invoke('prosop:analysis:workspace'),
+  runProsopAnalysis: (input) => ipcRenderer.invoke('prosop:analysis:run', input),
+  getProsopNetworksWorkspace: () => ipcRenderer.invoke('prosop:networks:workspace'),
+  saveProsopNetworkLayer: (input) => ipcRenderer.invoke('prosop:networks:saveLayer', input),
+  saveProsopNetworkEdge: (input) => ipcRenderer.invoke('prosop:networks:saveEdge', input),
+  deriveProsopCooccurrenceLayer: (layerId) => ipcRenderer.invoke('prosop:networks:deriveCooccurrence', layerId),
+  searchProsopography: (query, kind) => ipcRenderer.invoke('prosop:search', query, kind),
+  listProsopProposals: () => ipcRenderer.invoke('prosop:proposals:list'),
+  createProsopProposal: (input) => ipcRenderer.invoke('prosop:proposals:create', input),
+  decideProsopProposal: (proposalId, status, reviewedBy, decisionNote) => ipcRenderer.invoke('prosop:proposals:decide', proposalId, status, reviewedBy, decisionNote),
+  saveProsopNoteLink: (input) => ipcRenderer.invoke('prosop:notes:link', input),
+  exportProsopLongRows: () => ipcRenderer.invoke('prosop:export:long'),
+  exportProsopIpif: () => ipcRenderer.invoke('prosop:export:ipif'),
+  auditProsopIntegrity: () => ipcRenderer.invoke('prosop:integrity:audit'),
+  seedProsopDemo: () => ipcRenderer.invoke('prosop:demo:seed'),
+};
